tests: don't run Ginkgo tests in parallel
Ginkgo can run in parallel, but doing so spins up separate `go test`
processes and an operator for *each* worker / CPU core.

This creates competing operators that fight to process the same Stack CRs, and
causes concurrency issues that ultimately lead to indeterministic update states.

Spawning a single operator in Ginkgo to share amongst a set of tests
would be ideal, but Ginkgo does not support running shared services in a
global context for the entirety of the test suite.

Ultimately, the operator will need to be configured with leader election
to settle contention between multiple Operator instances. Once available,
this should allow ginkgo to run in parallel again.

See:
 - https://github.com/pulumi/pulumi-kubernetes-operator/issues/33
 - https://github.com/operator-framework/operator-sdk/issues/3585
 - https://onsi.github.io/ginkgo/#parallel-specs
 - https://docs.openshift.com/container-platform/4.5/operators/operator_sdk/osdk-leader-election.html