Add `backend` configuration (#106)
Adds a top level `backend` field to the Stack CR that allows pointing all operations for the target Stack at the given backend.  In cases where `backend` is set to a non-service backend, the AccessTokenSecret field is also no longer required.

This is dependent on https://github.com/pulumi/pulumi/pull/5789, and cannot be merged until the operator can take a dependency on that change in the core `pulumi` CLI.

Fixes #83.