Pick a single, watched namespace for leader lock
At present when the environment variable `WATCH_NAMESPACE` is given,
it is used verbatim as the namespace for the leadership election
module; but, the value can be a comma-delimited list, e.g., `foo,bar`,
which are not valid namespace names.

This commit fixes that by ensuring a single namespace name is
used. Any of the watched namespaced would do, on the assumption that
no other deployment of this operator will be using those namespaces.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>