Guard deref of pointer field .Status.LastUpdate
For a new resource, the status will be the zero value, meaning
pointer-valued fields will be nil. The controller dereferences the field
`.Status.LastUpdate` without first checking whether it's nil, leading to
a nil pointer dereference in this situation.

The quick solution in this case is to skip the offending block if the
field is nil. This avoids the nil pointer dereference; and it's correct
because the code within the block is irrelevant if there's no last
update.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>