Move deletion handling up front
At present, the git repo fields in the stack spec are validated only if
the resource is not marked to be deleted, but further work using those
fields is done unconditionally. This is incorrect, since invalid field
values might be used.

If the resource is due to be deleted, there is nothing else to do, so
the controller should just clean up and exit early (since the
finalizer-adding code goes with that, it moves too). Then the
validation, and further work if validation succeeds, can proceed
unconditionally.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>