Bump controller-gen to 0.9.2 (#314)
* Bump controller-gen to 0.9.2

controller-gen 0.5.0 is over a year old, and depends on a version of
golang/x/sys that doesn't compile with go 1.18.

This commit also includes cosmetic changes to files resulting from

    make codegen

* Reinstate additional printer columns

In #268, additional printer columns were added to the CRD file. However,
this file is generated, and regenerating it loses the extra stanza.

The sustainable way to do this is to use the kubebuilder annotation
`printcolumn`. (NB the column name is upper-cased by the UI; examples use
a proper-cased string.)

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>