Update .status and exit cleanly on stack failures
There's no need to return an error if the stack could not be processed
as defined -- that's a problem with the Stack object (or its
dependencies), not with the processing itself. In fact, it's bad to
return an error, since:

 - this will record a stack trace to the log, which is misleading
   because it suggests there's a bug; and,
 - the object will always be requeued; generally, we want requeues to be
   explicit.

This commit changes most failures such that they update the object
.status to reflect the failure, and requeue*, but do not return an
error.

(*In general you would expect some intervention to be necessary for a
stack to succeed after failing; however, most of these interventions
will be invisible to the controller, e.g., changes to a ConfigMap, or
something external to the cluster. So the best we can do is retry with
backoff.)

It also makes some minor adjustments so that saving the status is always
done on exit, rather than relying on whichever code happens to run
before a return. The observedGeneration is saved on a clean exit.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>