Rationalise object and object status updates
The previous scheme for updating the object or object status relied on
calling `getLatestResource()` before updating to avoid conflicts. But
this causes a problem, because it mutates the local version of the
object -- so the specification it is acting on can change
mid-flight.

For status, there are a handful of changes that are made during the
course of processing an object, and the tactic is to use a patch
operation. This applies the changes made locally, without stepping on
other status updates -- though at present there are no others expected,
other than those made previously by this controller.

For object updates, there's only one thing that gets updated -- adding
and removing the finalizer -- and this can be done by retrying an update
function which makes that exact change. Patching would also work, but
it's simpler to do the specific thing needed.

This commit also threads the context through to calls which use the
client, so that cancellations have a chance of working. This will help
shutdowns be more graceful (or at least, to be quicker).

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>