Pass ctx through to subroutines needing a context
This replaces uses of `context.Background` and `context.TODO` with a
context passed as an argument. Using `context.Background` meant that any
deadline or cancellation of the request would not be propagated (a
problem for graceful shutdown), and success or failure might go
unobserved (if the instigating request is abandoned but a remote process
is not). This kind of system will never be transactional, but we can at
least give it the best chance possible :-)

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>