Give better instructions for stale state solution
There's an ordering problem, as the instructions stand -- you don't
usually know the URN of a resource until the program has been run. So:
adapt the instructions for solving the stale provider state problem,
such that the "main" stack is run once, and the secondary stack is
created afterwards, once you know the URN in question.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>