Reduce individual task timeout
The timeout for uses of `Eventually` is 10 minutes, but a whole test
case times out after 10 minutes -- so any failures will 1. take 10
minutes, and 2. tend to trigger the more opaque whole test case timeout,
where you get a process dump rather than a test failure mentioning the
line number.

This commit reduces the timeout for indvidual `Eventually` calls to 1
minute -- if this turns out to be optimistic in specific cases, some can
be increased (or it can be taken as a signal that the test should be
reconsidered).

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>