Test that prerequisites+targets solve the problem
This adds a test that uses prerequisites and targets together to solve
the stale provider state problem.

It is difficult at present to show that it is specifically those
elements that make it work -- it could work by coincidence, since if the
controller keeps retrying, eventually the prerequisite stack will have
run whether it was referenced as a prerequisite or not. The means I've
used here is to let the prerequisite stack succeed once, so it's not in
a retry loop, _then_ invalidate the provider state by restarting
RabbitMQ with a new password. Unless there's some mechanism to provoke
the prerequisite stack into running again, the stack using the stale
state cannot succeed.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>