Implement reconciliation request mechanism
This adds a mechanism for "requesting" the reconciliation of a
particular object, which gives a way for prerequisites to be queued when
they haven't been processed recently enough.

The mechanism itself is simple: annotate the object in question, so that
the Kubernetes API sends an update to watchers. The controller needs to
pay attention to these changes, which entails a change in predicates:
instead of "the generation has changed OR there is no generation*", it
is "the generation changed or the annotation has been added or amended".

Finally, if the object was reconciled without crashing, the annotation
value is saved into the status. This is so that tooling can see it was
acted upon.

*Stack objects do have a generation, so there is no reason for this
predicate.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>