Parameterise the namespace(s) and versions
This commit makes it possible to install the operator in several
namespaces, by supplying multiple values in the config item
`namespaces` when running one of the Pulumi programs under deploy/.

(This also makes the deployed operator log level `error` in the Go and
.NET programs, to bring them in line with the other programs and
canonical YAML configuration.)

The CRD and operator tags are now items of config, for each of the
programs under deploy/, with defaults. An upgrade is then a case of
e.g.,

    pulumi config set operator-version v1.10.2
    pulumi up

The defaults are consts at the top of the file, and can be updated more
easily.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>