Use conditions to indicate ready state (#316)
* Fix accidental use of err from outer scope

* Give Stack.status an observedGeneration field

The observedGeneration field is a convention used to distinguish between
being marked ready because the object is up to date, or because changes
have not been seen yet.

* Update .status and exit cleanly on stack failures

There's no need to return an error if the stack could not be processed
as defined -- that's a problem with the Stack object (or its
dependencies), not with the processing itself. In fact, it's bad to
return an error, since:

 - this will record a stack trace to the log, which is misleading
   because it suggests there's a bug; and,
 - the object will always be requeued; generally, we want requeues to be
   explicit.

This commit changes most failures such that they update the object
.status to reflect the failure, and requeue*, but do not return an
error.

(*In general you would expect some intervention to be necessary for a
stack to succeed after failing; however, most of these interventions
will be invisible to the controller, e.g., changes to a ConfigMap, or
something external to the cluster. So the best we can do is retry with
backoff.)

It also makes some minor adjustments so that saving the status is always
done on exit, rather than relying on whichever code happens to run
before a return. The observedGeneration is saved on a clean exit.

* Be consistent about which failures are crashes

This changes the remainder of the places where a "crash" was signalled
(something unexpected happened!) but it's actually a problem that you'd
expect to see in the normal course of things, like the spec being
invalid. In some cases, the problem might just go away, so retrying is
worthwhile.

* Add conditions to Stack.status

Conventionally, conditions are used in Kubernetes resources to report
observations of the state of processing. For our purposes, they will
help people and tooling understand whether a particular Stack has been
run as specified successfully.

NB., since older API versions don't include conditions, the definition
of StackStatus is no longer shared.

This commit also documents the intended scheme for using conditions.

* Rationalise object and object status updates

The previous scheme for updating the object or object status relied on
calling `getLatestResource()` before updating to avoid conflicts. But
this causes a problem, because it mutates the local version of the
object -- so the specification it is acting on can change
mid-flight.

For status, there are a handful of changes that are made during the
course of processing an object, and the tactic is to use a patch
operation. This applies the changes made locally, without stepping on
other status updates -- though at present there are no others expected,
other than those made previously by this controller.

For object updates, there's only one thing that gets updated -- adding
and removing the finalizer -- and this can be done by retrying an update
function which makes that exact change. Patching would also work, but
it's simpler to do the specific thing needed.

This commit also threads the context through to calls which use the
client, so that cancellations have a chance of working. This will help
shutdowns be more graceful (or at least, to be quicker).

* Pass ctx through to subroutines needing a context

This replaces uses of `context.Background` and `context.TODO` with a
context passed as an argument. Using `context.Background` meant that any
deadline or cancellation of the request would not be propagated (a
problem for graceful shutdown), and success or failure might go
unobserved (if the instigating request is abandoned but a remote process
is not). This kind of system will never be transactional, but we can at
least give it the best chance possible :-)

* Maintain status conditions within Reconcile

This makes the controller follow the "ready protocol", by updating an
object's status conditions as it is processed.

* Changelog entry for adding status conditions

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>