Deprecate cross-namespace refs
Cross-namespace refs won't work with the provided configuration, which
gives RBAC permissions for one namespace only. Granting permissions to
other namespaces means they are also accessible to the programs run by
stacks, which is a security hole.

Therefore: guard the use of `api/pulumi/shared/SecretRef.Namespace`, and
watching >1 (or all) namespaces with `WATCH_NAMESPACE`, with an
environment variable `INSECURE_NO_NAMESPACE_ISOLATION`. This must be set
to `1` or `true` to enable cross-namespace operation.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>