Use request mechanism to queue stale prerequisites
Previous commits added

 - an API for specifying prerequisites for a Stack (that is, other
   stacks which must be up to date)
 - a mechanism for annotating a Stack so it will get re-run

This commit ties those together by checking each prerequisite, before
running a stack, and queueing any that are out of date.

It works but is inefficient, because it requeues a Stack with backoff,
which means it will likely fail several more times before its
prerequisites have a chance to run (backoff starts on `O(1ms)`, and
typical stack running times are `O(10s)`).

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>