Use watch to requeue stack dependents
When a stack has out-of-date prerequisites, it needs to queue those and
wait until they have succeeded before trying again. Because of the way
controller-runtime works, we can't wait for success exactly, but we
_can_ react to the prerequisites being updated by re-evaluating
requirements.

This commit also rearranges the tests a little to avoid a container
restart, so to run a little quicker.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>