Make a cross-namespace ref cause a stalled state
If you have a cross-namespace ref in your Stack, and namespace isolation
has not been waived, you will need to change the stack or reconfigure
the controller in order to make progress. This commit introduces a tiny
protocol (that could be used throughout, hint hint) so a helper func can
signal that an error should stall the object, rather than letting it be
retried; and, uses it in `resolveResourceRef` so that forbidden ref will
result in a stalled state.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>