Parallelise and randomize tests (#331)
Many of the tests take O(1m) to run, and the more specs added the longer
the runtime. Parallelising the tests will help defer the day we need to
increase the overall timeout.

Randomising helps keep the tests independent, by surfacing things that
work by coincidence.

Enabling either of these requires the use of the Ginkgo command line --
it's not possible to enable them programmatically. So: reinstate the use
of `ginkgo` for running tests.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>