Tidy git repo source fields into an inlined struct
To prepare the way for referring to other sources of Plumui programs in
the Stack spec, this commit moves the fields specifying how to fetch
from a git repo directly into their own struct. It's not possible to
change the _schema_ in a backward-compatible way, so the struct is
serialised inline and won't look any different in the API.

There is a semantic change, though: the field `.spec.projectRepo`
becomes optional (the others involved were already optional).

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>