Deprecate cross-namespace refs and watching multiple namespaces (#329)
* Deprecate cross-namespace refs

Cross-namespace refs won't work with the provided configuration, which
gives RBAC permissions for one namespace only. Granting permissions to
other namespaces means they are also accessible to the programs run by
stacks, which is a security hole.

Therefore: guard the use of `api/pulumi/shared/SecretRef.Namespace`, and
watching >1 (or all) namespaces with `WATCH_NAMESPACE`, with an
environment variable `INSECURE_NO_NAMESPACE_ISOLATION`. This must be set
to `1` or `true` to enable cross-namespace operation.

* Mention deprecation in changelog

* Replace github.com/pkg/errors with errors and fmt

errors.Wrap is used extensively to provide extra context to errors. The
Go standard library now builds in wrapped errors, with `fmt.Errorf` and
the formatting instruction `%w`. This commit replaces
github.com/pkg/errors (which is archived) with uses of `fmt.Errorf`.

* Test that cross-ns refs are forbidden

 - show that without INSECURE_NO_NAMESPACE_ISOLATION=true, the
   controller will fail a stack that uses a cross-ns ref
 - show that with the env entry, it succeeds

* Make a cross-namespace ref cause a stalled state

If you have a cross-namespace ref in your Stack, and namespace isolation
has not been waived, you will need to change the stack or reconfigure
the controller in order to make progress. This commit introduces a tiny
protocol (that could be used throughout, hint hint) so a helper func can
signal that an error should stall the object, rather than letting it be
retried; and, uses it in `resolveResourceRef` so that forbidden ref will
result in a stalled state.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>