Adapt the controller to cope with missing git spec
Having made .projectRepo optional, leaving all fields in the inlined
GitSource optional, we now have to deal with the possibility of it being
missing entirely.

NB since the git repo needs to be cloned even if the stack is being
deleted, the validation of branch|commit is necessary either way (this
was the case before, it's just more obvious now I've moved the check
later).

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>