Port to Ginkgo v2
This updates Ginkgo, the test frameworks, to v2. It is largely the same,
except 1. custom reporters are even more deprecated (I removed the code
which set up a special test reporter; this should get rid of the
deprecation message); 2. we can now use DeferCleanup, which is a clearer
than having "matching" code in BeforeEach/AfterEach.

I've rewritten the cleanup code which used `toDelete []string`. Because
the invocation was outside a BeforeEach/It body, it executed during the
tree construction phase[1] rather than when the tests were run, so it
didn't actually clean anything up.

[1]: https://onsi.github.io/ginkgo/#mental-model-how-ginkgo-traverses-the-spec-hierarchy

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>