Avoid race between waitForStack* and reprocessing
Many of the tests wait for a particular value of
`.status.lastUpdate.State`, then check the values of conditions are as
expected. However, between these two things, they fetch the object from
the API server again (`refetch`), since the waitForStack* uses a
temporary variable rather than mutating the stack given to it. That that
admits the possibility that the status might have been updated again, by
the controller reconciling it again.

Instead, let waitForStack* write to the object it's given, and don't
`refetch` -- then, the status observed will be the status at the point
the wait succeeded.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>