Insist on exactly one source spec
Previously, an if-then-else checked for the presence of source
specifications (GitSource and FluxSource), and this made it possible to
supply a stack object with both kinds of source (though only the first
checked -- GitSource -- would be used).

This commit makes that bit of code enforce the mutual exclusion of
sources.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>