Make mutual exclusion of sources obvious
This changes the which-source-is-it switch around a little, so that
instead of each case being "this is non-nil and all the others are nil",
the first case checks that exactly one is non-nil. This makes it a
little easier to read and extend; you only have to update the constraint
at the top, then add your particular source in as a case.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>