Remove "catch all" deletion of test stacks
The cleanup procedure for tests in stack_controller_test.go deletes all
stacks in the namespace, then waits for them to disappear from the API
results (i.e., they have been finalised and removed from the
cache). This has the problem that it can foul on other tests, if they
don't tidy up their own stacks. It will catch when a test case has
simply neglected to delete a stack it created, but not when a stack
fails to be finalised, and in the latter case, it takes a long time to
fail, and is difficult to connect to the actual problem.

On balance I think it's better to remove a source of mysterious
failures, at the expense of less chance of noticing if a test case leaks
resources.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>