Tease apart stack_controller_test.go test cases
This file has what are more or less end-to-end tests, which are few and
each quite long-running. Here I have rearranged the tests so that set-up
and tear-down is clearer, and so they are putting fewer things to the
test at a time (and there is further this can go).

I have also made most of the tests use the file backend rather than the
service, in the expectation that they will complete faster. And, rather
than using the v1alpha1 and v1 types throughout, I've made an explicit
test covering their compatibility, and used v1 everywhere else.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>