Shortcut for setting the number of test nodes
I find that on my laptop, running two test nodes (each one a control
plane instance of `{etcd, kube-apiserver}`) is more reliable than trying
for four. Sharing the control plane and isolating with namespaces would
help, but in the meantime to make this a bit more convenient, this
commit makes the number of nodes a variable in the Makefile.

    make test TEST_NODES=2

(.. and, incidentally, --randomize-all is the Ginkgo v2 spelling of
--randomizeAllSpecs)

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>