Remove race condition from "retrying" test
If the controller repeatedly retries a Stack, it will oscillate between
having the "Reconciling" condition because it has been requeued, and
having the same condition because it's being processed by the controller
again.

This makes seeing a specific reason for retrying tricky -- but, if
you've observed that the stack has been processed and failed, then
_either_ reason is good enough to show that it's being retried, so that
race condition can be avoided.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>