Expose failures to clean up after tests (#342)
* Expose failures to clean up after tests

When stack objects aren't deleted after tests, it can mean

 - resources created by the stack aren't tidied up;
 - the finalizer logic isn't well tested
 - the stack might interfere in annoying ways with other tests.

Previously, all stacks in the namespace were deleted as part of a
top-level `AfterEach`; but because this tried to delete all stacks and
wait until they'd been finalised, any rogue test case that didn't clean
up properly and couldn't be finalised would cause some _other_ test to
time out.

Instead, this commit uses a similar idea, but fails the test suite if
there are stacks left over at the very end. This highlighted a couple of
problems, which I've also fixed here:

 - one test was deleting its working directory before deleting the stack
   using it, and would fail to be finalized;
 - lots of stacks had random names that didn't help with tracking down
   the test case

Finally, the helper `deleteAndWaitForFinalization` can be used to ensure
a test case doesn't exit until it has truly cleaned up. It's worth using
this everywhere -- next, I'm going to rewrite tests to use it, and the
other helpers.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>

* Use deletion helper throughout

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>

* Replace testdata/empty-stack with YAML program

It's a little bit faster to run a YAML program than to download
dependencies and run a NodeJS program, so this saves a bit of time and
reduces the possibility of timeouts.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>

* Raise the slow test threshold

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>

* Remove race condition from "retrying" test

If the controller repeatedly retries a Stack, it will oscillate between
having the "Reconciling" condition because it has been requeued, and
having the same condition because it's being processed by the controller
again.

This makes seeing a specific reason for retrying tricky -- but, if
you've observed that the stack has been processed and failed, then
_either_ reason is good enough to show that it's being retried, so that
race condition can be avoided.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>