Let ContinueResyncOnCommitMatch apply to all sources (#346)
* Move ContinueResyncOnCommitMatch out of *GitSource

This field is relevant to other sources (FluxSource and ProgramRef), so
deserves to be in .Spec rather than the embedded *GitSource.

Speaking of that latter, the controller code accesses
`.Spec.Continue[...]Match` even when *GitSource is nil, but only if the
status has been updated once. This results in a nil pointer
dereference. Moving the field into .Spec eliminates that problem.

* Log entry in changelog

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>