Check for an artifact in Flux sources
When a Stack refers to a Flux source, we check the latter for a `Ready`
condition, and if present and false, bail on reconciliation until it
gets requeued by the watch mechanism.

A flaw in this is that a new source will have no Ready condition (true
or false), so the controller will continue, only to find there's no
record of what it should download. Instead of gracefully parking itself,
it bounces off this until the source has been seen by the
source-controller.

So: check that there's an artifact.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>